fix(review): finish the extraction — tests, visibility, formatting
Review round 1 on the registry/status split.

Tests now really do follow their subjects. The first pass moved one test
and left eleven behind in `api/orders.rs`, which is why the bridge file
still owned the registry's coverage. Moved: the eight `take_matching_*` /
detach / concurrency / classify tests and the id-reconciliation and
content-key lookup ones to `mostro::pending`; the wire-status precedence,
hard-terminal, add-invoice-sync, escrow-locked and cancellation-wipe ones
to `mostro::status`. `api/orders.rs` 4515 → 3911.

Three tests written in the first pass turned out to restate rules the
moved ones already pin more thoroughly, so they are gone rather than
duplicated: the rebind rule, the wire-status precedence, and the
cancellation wipe. `small_order_with` is needed by both modules now, so
it lives in a shared `mostro::test_fixtures` instead of being spelled out
twice.

Visibility: `PENDING_REQUESTS`, `request_id_matches` and
`is_terminal_status` had no caller outside their own module and are
private again — widening them worked against the encapsulation this
refactor exists to create. Both modules are `pub(crate) mod`, matching
the fact that every item in them is `pub(crate)`.

Also: rustfmt on the new files (the `pub(crate)` prefix had pushed lines
past 100 columns), the doc comment on `map_core_status` reattached to
`map_core_status` instead of dangling on `status_for_action`, a module
doc saying which of `mostro::status` and `mostro::fsm` owns what, and a
let-unit binding clippy flagged.

Rebased on main, which carried the gift-wrap → transport rename through
the moved code: the debug tag is `daemon-msg` and the docs no longer say
gift-wrap.

cargo test 257 passed (260 minus the 3 duplicates), clippy clean on every
touched file, flutter analyze 7 pre-existing, flutter test 229 passed,
frb-generate --check clean.
