+++ rust/src/mostro/mod.rs
@@ -1,11 +1,13 @@
 pub mod actions;
 pub mod escrow_mode;
 pub mod fsm;
-pub mod pending;
+pub(crate) mod pending;
 pub mod pow;
 pub mod protocol_version;
 pub mod session;
-pub mod status;
+pub(crate) mod status;
 
 #[cfg(test)]
 mod cashu_wire;
+#[cfg(test)]
+pub(crate) mod test_fixtures;
